+++ meteorhub/src/components/SearchComponent.jsx
@@ -1,39 +1,22 @@
-import React, { useState, useEffect } from "react";
 import "./SearchComponent.css";
 import SummaryMetric from "./SummaryMetric";
 import DataDisplayTable from "./DataDisplayTable";
 
-export default function SearchComponent() {
-  const [data, setData] = useState([]);
-  const [query, setQuery] = useState("");
-  const [filteredData, setFilteredData] = useState([]);
-  useEffect(() => {
-    fetch("./src/assets/data.json")
-      .then((response) => response.json())
-      .then((jsonData) => {
-        setData(jsonData);
-        setFilteredData(jsonData);
-      })
-      .catch((error) => console.error("Error fetching data:", error));
-  }, []);
-  useEffect(() => {
-    const filteredResults = data.filter(
-      (item) =>
-        item.name.toLowerCase().includes(query.toLowerCase()) ||
-        item.year.toLowerCase().includes(query.toLowerCase()) ||
-        item.recclass.toLowerCase().includes(query.toLowerCase())
-    );
-    setFilteredData(filteredResults);
-  }, [data, query]);
+export default function SearchComponent({
+  query,
+  filteredData,
+  onChangeHandler,
+  onClickHandler,
+}) {
   return (
     <>
       <input
         type="text"
         placeholder="Search"
         value={query}
-        onChange={(e) => setQuery(e.target.value)}
+        onChange={(e) => onChangeHandler(e)}
       />
-      <button className="clear" onClick={(e) => setQuery("")}>
+      <button className="clear" onClick={(e) => onClickHandler(e)}>
         Clear
       </button>
       <SummaryMetric filteredData={filteredData} />
